+++ code/js/browser-action.js
@@ -2,6 +2,8 @@
 
 console.log('loaded: browser-action.js')
 
+let bypassCache = false;
+
 // browser.browser​Action
 // https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/browserAction
 // A browser action is a button in the browser's toolbar.
@@ -36,7 +38,17 @@ console.log('loaded: browser-action.js')
 browser.browserAction.onClicked.addListener(
   //
   // tab: tabs.Tab -> The Tab that was active when the icon was clicked.
-  function (tab) {
+  // onClickData -> An object containing information about the click
+  //  {
+  //    modifiers -> An array. The keyboard modifiers active at the time of the click,
+  //                 being one or more of Shift, Alt, Command, Ctrl, or MacCtrl.
+  //    button -> An integer. Indicates the button used to click the page action icon: 0 for
+  //              a left - click or a click not associated with a mouse, such as one from the
+  //              keyboard and 1 for a middle button or wheel click. Note that the right-click
+  //              is not supported because Firefox consumes that click to display the context
+  //              menu before this event is triggered.
+  //  }
+  function (tab, OnClickData) {
     //
     // tabs.query
     // https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/tabs/query
@@ -54,17 +66,28 @@ browser.browserAction.onClicked.addListener(
     //   windowId: integer -> The ID of the parent window, or windows.WINDOW_ID_CURRENT for the
     //                        current window.
     // }
-    browser.tabs
-      .query({ windowId: tab.windowId })
-      .then(reload_each_tab_in_list)
-      .catch(err => {
-        console.log('[browser-action.js@tabs.query] error:', err)
+    browser.storage.local
+      .get('bypass-cache-by-default')
+      .then(result => {
+        setBypassCache(result, OnClickData.modifiers)
+        browser.tabs
+          .query({ windowId: tab.windowId })
+          .then(reload_each_tab_in_list)
+          .catch(err => {
+            console.log('[browser-action.js@tabs.query] error:', err)
+          })
       })
+      .catch(err => console.log('[browser-action.js@reload] error:', err))
   }
 )
 
+function setBypassCache(storage, modifiers) {
+  let storage_value = (typeof storage['bypass-cache-by-default'] === 'undefined') ? false : storage['bypass-cache-by-default']
+  bypassCache = storage_value ? !modifiers.includes('Shift') : modifiers.includes('Shift')
+}
+
 // tabList: Array
-function reload_each_tab_in_list (tabList) {
+function reload_each_tab_in_list(tabList) {
   tabList.filter(tabs_to_reload).forEach(reload)
 }
 
@@ -83,15 +106,23 @@ const reload = function (tab) {
   // Reload a tab, optionally bypassing the local web cache.
   //
   // tab.id: integer -> The ID of the tab to reload.
-  browser.tabs.reload(tab.id)
-  console.log('[browser-action.js] info: reloaded tab', tab.id)
+  // reloadProperties: object
+  // {
+  //   bypassCache: boolean -> Bypass the local web cache. Default is false.
+  // }
+
+  browser.tabs.reload(tab.id, { bypassCache })
+  if (bypassCache)
+    console.log('[browser-action.js] info: [cache-disabled] reloaded tab', tab.id)
+  else
+    console.log('[browser-action.js] info: reloaded tab', tab.id)
 }
 
 // Set of tab ids to be ignored when reloading tabs in window.
 const ignore_list = new Set()
 
 // tabId: integer
-function ignore_list_add (tabId) {
+function ignore_list_add(tabId) {
   if (!ignore_list.has(tabId)) {
     ignore_list.add(tabId)
     //
@@ -106,7 +137,7 @@ function ignore_list_add (tabId) {
 }
 
 // tabId: integer
-function ignore_list_remove (tabId) {
+function ignore_list_remove(tabId) {
   if (ignore_list.has(tabId)) {
     ignore_list.delete(tabId)
     //
@@ -120,7 +151,7 @@ function ignore_list_remove (tabId) {
 }
 
 // tabId: integer
-function is_ignored (tabId) {
+function is_ignored(tabId) {
   return ignore_list.has(tabId)
 }